refactor: ProblemService 코드 개선
