+++ src/main/java/com/Alchive/backend/controller/ProblemController.java
@@ -3,7 +3,6 @@ package com.Alchive.backend.controller;
 import com.Alchive.backend.config.Code;
 import com.Alchive.backend.config.exception.NoSuchProblemException;
 import com.Alchive.backend.dto.request.ProblemCreateRequest;
-import com.Alchive.backend.dto.request.SubmitProblemCreateRequest;
 import com.Alchive.backend.dto.response.ApiResponse;
 import com.Alchive.backend.dto.response.ProblemDetailResponseDTO;
 import com.Alchive.backend.dto.response.ProblemListResponseDTO;
@@ -14,8 +13,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
@@ -44,7 +41,7 @@ public class ProblemController {
     @PostMapping("/submit")
     public ResponseEntity<ApiResponse> createProblemSubmit(
             HttpServletRequest tokenRequest,
-            @RequestBody @Valid SubmitProblemCreateRequest problemRequest) {
+            @RequestBody @Valid ProblemCreateRequest problemRequest) {
         problemService.createProblemSubmit(tokenRequest, problemRequest);
         return ResponseEntity.ok()
                 .body(new ApiResponse(HttpStatus.OK.value(), "제출한 문제와 코드 정보를 저장했습니다."));
