+++ src/main/java/com/Alchive/backend/dto/response/ProblemListResponseDTO.java
@@ -1,6 +1,7 @@
 package com.Alchive.backend.dto.response;
 
 import com.Alchive.backend.domain.Algorithm;
+import com.Alchive.backend.domain.Problem;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -18,4 +19,15 @@ public class ProblemListResponseDTO {
     private String problemState;
     private List<Algorithm> algorithmList;
 
+    public static ProblemListResponseDTO of(Problem problem, List<Algorithm> algorithmList) {
+        return ProblemListResponseDTO.builder()
+                .problemId(problem.getProblemId())
+                .problemNumber(problem.getProblemNumber())
+                .problemTitle(problem.getProblemTitle())
+                .problemDifficulty(problem.getProblemDifficulty())
+                .problemPlatform(problem.getProblemPlatform())
+                .problemState(problem.getProblemState())
+                .algorithmList(algorithmList)
+                .build();
+    }
 }
